fix: Only switch dir for building with preset
config:build_directory_path() is relative to nvim's cwd, not the cwd as
given in the config, so cmake --build will fail if config.cwd is a
subdirectory of nvim's cwd. On the other hand, when using build presets
CMake's cwd *must* be the directory where the preset files are. Fix this
by only using config.cwd when building with presets.